Collect almost all the links together
The links are now (mostly) organised by topic at the end of
assorted-text.txt, except for an "unsorted" section at the
end of the same document. This isn't just making lists for the sake of
it, I actually want to use this in helping to write the main section for
each topic in the final extended notes.