Take out 2 more slides from the short slide set
This *should* make it more like a 25 minute presentation.

The long slide set is 22 slides (21 excluding Fin), and the short
one is now 16 (15 excluding Fin). 15/21 is 0.7, which is a bit
optimistic given 25/45 is 0.5 recurring, but given the previous
version of the short set at PyCon UK took about 30 minutes,
removing 2 slides should do it...